backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-anthropic
langchain-openai
langchain-google-genai
google-generativeai
pydantic
python-dotenv
httpx
tenacity
langgraph-checkpoint-sqlite
pyyaml
